code/requirements.txt
numpy>=1.23
pandas>=2.0
pyyaml>=6.0
tqdm>=4.66
scikit-learn>=1.2
matplotlib>=3.7
networkx>=3.1

torch>=2.0
torch-geometric>=2.3
